refactor(core-weave): extract reference catalog link parsing
- move current ReferenceCatalog link parsing into reference_catalog_links.ts
- keep role-label derivation private to the parser module
- preserve planner dispatch, generated RDF/Page output, and public weave façade exports
